Locking down delete item permissions

+++ backend/src/resolvers/Mutation.js
@@ -25,7 +25,12 @@ const Mutation = {
     },
     async deleteItem(parent, args, ctx, info) {
         const where = { id: args.id };
-        await ctx.db.query.item({ where }, `{ id, title }`);
+        const item = await ctx.db.query.item({ where }, `{ id, title user { id }}`);
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some(
+            permission => ['ADMIN', 'ITEMDELETE'].includes(permission)
+        );
+        if (!ownsItem && !hasPermissions) throw new Error('You don\'t have permissions to do that');
         return ctx.db.mutation.deleteItem({ where }, info);
     },
     async signup(parent, args, ctx, info) {
